refactor: errormessage file renamed to toastmessage
